# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/security-schemes": {
      "post": {
        "operationId": "SecuritySchemesController_create",
        "summary": "Create a reusable issue security scheme (Module 6)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSecuritySchemeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "security-schemes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "SecuritySchemesController_list",
        "summary": "List this organization's security schemes",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "security-schemes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/security-schemes/{id}": {
      "patch": {
        "operationId": "SecuritySchemesController_update",
        "summary": "Update a security scheme's name and/or levels",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSecuritySchemeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "security-schemes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "SecuritySchemesController_remove",
        "summary": "Delete a security scheme (must not be assigned to any project)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "security-schemes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/teams": {
      "post": {
        "operationId": "TeamsController_create",
        "summary": "Create a team (Module 6: Teams, Project Roles & Security Schemes)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "teams"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "TeamsController_list",
        "summary": "List this organization's teams",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "teams"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/teams/{id}": {
      "get": {
        "operationId": "TeamsController_get",
        "summary": "Get a single team",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "teams"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "TeamsController_update",
        "summary": "Update a team's name, description, lead, and/or members",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "teams"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "TeamsController_remove",
        "summary": "Delete a team",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "teams"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Project & Task Management API",
    "description": "REST API for a Jira/Trello-style project and task management system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SecurityLevelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Confidential"
          },
          "allowedRoles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "PlatformAdmin",
                "Admin",
                "Manager",
                "Developer"
              ]
            }
          },
          "allowedUserIds": {
            "description": "User ids individually granted view access",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowedTeamIds": {
            "description": "Team ids granted view access (Module 6)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowedProjectRoleIds": {
            "description": "Project Role ids granted view access (Module 6)",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "allowedRoles",
          "allowedUserIds"
        ]
      },
      "CreateSecuritySchemeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Standard confidentiality levels"
          },
          "levels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SecurityLevelDto"
            }
          }
        },
        "required": [
          "name",
          "levels"
        ]
      },
      "UpdateSecuritySchemeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Standard confidentiality levels"
          },
          "levels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SecurityLevelDto"
            }
          }
        }
      },
      "CreateTeamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Backend Guild"
          },
          "description": {
            "type": "string"
          },
          "leadId": {
            "type": "object",
            "nullable": true
          },
          "memberIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateTeamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Backend Guild"
          },
          "description": {
            "type": "string"
          },
          "leadId": {
            "type": "object",
            "nullable": true
          },
          "memberIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}